+++ addOns/pscanrules/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Maintenance changes.
+- Modern scan rule to refer to the Client Spider.
 
 ## [74] - 2026-06-26
 ### Changed